fix: keep feedback evaluations browser-safe
