+++ packages/docs/src/index.ts
@@ -90,10 +90,6 @@ export {
   DOCS_AGENT_FEEDBACK_EVALUATION_BASELINE_FORMAT,
   DOCS_AGENT_FEEDBACK_EVALUATION_CANDIDATES_FORMAT,
   DOCS_AGENT_FEEDBACK_EVALUATION_REGRESSION_FORMAT,
-  readDocsAgentEvaluationTasksFile,
-  readDocsAgentFeedbackEvaluationBaseline,
-  readDocsAgentFeedbackImprovementReport,
-  readDocsGoldenTasksReportFile,
 } from "./agent-feedback-evaluations.js";
 export type {
   BuildDocsAgentFeedbackEvaluationCandidatesOptions,
